+++ src/SpeedTrapListener.php
@@ -49,6 +49,14 @@ class SpeedTrapListener implements TestListener
     protected $reportLength;
 
     /**
+     * Whether the test runner should halt running additional tests after
+     * finding a slow test.
+     *
+     * @var bool
+     */
+    protected $stopOnSlow;
+
+    /**
      * Collection of slow tests.
      * Keys (string) => Printable label describing the test
      * Values (int) => Test execution time, in milliseconds
@@ -132,6 +140,10 @@ class SpeedTrapListener implements TestListener
         $label = $this->makeLabel($test);
 
         $this->slow[$label] = $time;
+
+        if ($this->stopOnSlow) {
+            $test->getTestResultObject()->stop();
+        }
     }
 
     /**
@@ -224,6 +236,7 @@ class SpeedTrapListener implements TestListener
     {
         $this->slowThreshold = $options['slowThreshold'] ?? 500;
         $this->reportLength = $options['reportLength'] ?? 10;
+        $this->stopOnSlow = $options['stopOnSlow'] ?? false;
     }
 
     /**